Move inflate_job to doa module
